+++ packages/ricos-content/web/src/normalization/Fixtures/textWrapContentStateExpected.ts
@@ -270,7 +270,6 @@ export default {
           alignment: 'center',
           size: 'large',
           width: 'full-width',
-          enableVoteRole: false,
           textWrap: 'wrap',
         },
       },